requirements.txt
opencv-python==4.2.0.34
Pillow==9.0.1
cryptography==44.0.0
torch==2.2.2
torchaudio==2.6.0
torchvision==0.17.2
plyer==2.1.0
pywin32==308
facenet_pytorch==2.6.0
matplotlib==3.9.4